+++ src/main/java/com/github/jlangch/venice/impl/util/StringUtil.java
@@ -548,9 +548,21 @@ public class StringUtil {
         }
     }
 
+    public static String padLeft(final String s, final int len) {
+        final int padLen = len - s.length();
+        return padLen > 0 ? repeat(' ', padLen) + s : s;
+    }
+
     public static String padRight(final String s, final int len) {
         final int padLen = len - s.length();
-        return padLen > 0 ? s + StringUtil.repeat(' ', padLen) : s;
+        return padLen > 0 ? s + repeat(' ', padLen) : s;
+    }
+
+    public static String padCenter(final String s, final int len) {
+        final int padLen = len - s.length();
+        final int padRight = padLen / 2;
+        final int padLeft = padLen - padRight;
+        return padLen > 0 ? repeat(' ', padLeft) + s  + repeat(' ', padRight) : s;
     }
 
     public static String toEscapedUnicode(final char ch) {